dense_map: Use 32-bit quantities for size and capacity
Advent of Code-scale problems will very likely never require hash tables
with more than 4,294,967,295 elements; might as well save a little bit
of space since some solutions make use of dense_sets nested inside of
dense_maps.

+++ dense_map.h
@@ -117,9 +117,9 @@ public:
 private:
     std::unique_ptr<bucket[]> buckets_;
     std::unique_ptr<uint8_t[]> states_; // TODO: pack?
-    size_t capacity_;
-    size_t size_;
-    size_t size_with_tombs_;
+    uint32_t capacity_;
+    uint32_t size_;
+    uint32_t size_with_tombs_;
     hasher hash_;
     key_equal equal_;
